+++ src/views/user-list/index.vue
@@ -5,7 +5,7 @@ import { watch, onBeforeMount, ref, StyleValue } from 'vue';
 import { useEasyToggle } from '@/composables/use-easy-toggle';
 import { IHttpOptions, useHttp } from '@/utils/http';
 import { User } from '@/types/user';
-import UserCard from '@/components/user-card/index.vue';
+import UserCard from '@/views/user-list/components/user-card/index.vue';
 import { useUserStore } from '@/stores/user';
 import { usePopover } from '@/composables/use-popover';
 import UserForm from '@/components/user-form/user-form.vue';
